Elousa me ta commits

+++ src/main/java/com/coredumped/skyroof_constructions/dao/ProjectDao.java
@@ -3,5 +3,5 @@ package com.coredumped.skyroof_constructions.dao;
 import com.coredumped.skyroof_constructions.model.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectDao extends JpaRepository <Project,Integer> {
+public interface ProjectDao extends JpaRepository <Project,Long> {
 }
